pyproject.toml
[project]
name = "spotim8"
version = "6.0.0"
description = "Spotify library sync and yearly archive playlists. Tidy DataFrames, CLI, and streaming history."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "spotipy>=2.23",
  "pyarrow>=14; platform_system != 'Windows'",
  "tqdm>=4.66",
  "requests>=2.31",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.6", "black>=24.0"]
notebook = ["jupyter>=1.0", "matplotlib>=3.8", "seaborn>=0.13", "plotly>=5.18"]
all = ["spotim8[dev,notebook]"]

[project.scripts]
spotim8 = "src.core.cli:main"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
